requirements.txt
Django
djangorestframework
djangorestframework-jwt
django-allauth
djoser
drf-yasg
